// src/app/features/function-permission-group-maintenance/pages/user-settings/models/mgr-system-user-query-response.model.ts
export interface MgrSystemUserDataQueryResponse {
    DBSource: string;
    UserId: string;
    UserName: string;
    FuctionGroup: string;
    Email: string;
    Role: string;
    ActivityStartDate: string;
    ActivityEndDate: string;
    Status: string;
    LoginErrors: string;
    ModDate: string;
    ModTime: string;
    ModUser: string;
}